+++ plog.go
@@ -60,6 +60,16 @@ func Fatalf(format string, err error) {
 	loggers.write(newLogf(FatalLevel, format, err))
 }
 
+// TFatal will print a fatal error message and meta-tag the log
+func TFatal(tags Tags, err error) {
+	loggers.write(newTLogf(FatalLevel, tags, "%v", err))
+}
+
+// TFatalf will print a formatted, fatal error message and meta-tag the log
+func TFatalf(tags Tags, format string, err error) {
+	loggers.write(newTLogf(FatalLevel, tags, format, err))
+}
+
 //
 // Error logging (Level 2)
 //
@@ -72,6 +82,16 @@ func Error(err error) {
 // Errorf will print a formatted, non-fatal error message
 func Errorf(format string, err error) {
 	loggers.write(newLogf(ErrorLevel, format, err))
+}
+
+// TError will print a non-fatal error message and meta-tag the log
+func TError(tags Tags, err error) {
+	loggers.write(newTLogf(ErrorLevel, tags, "%v", err))
+}
+
+// TErrorf will print a formatted, non-fatal error message and meta-tag the log
+func TErrorf(tags Tags, format string, err error) {
+	loggers.write(newTLogf(ErrorLevel, tags, format, err))
 }
 
 //
@@ -88,6 +108,16 @@ func Warnf(format string, variables ...interface{}) {
 	loggers.write(newLogf(WarnLevel, format, variables...))
 }
 
+// TWarn will print any number of variables at warn level and meta-tag the log
+func TWarn(tags Tags, variables ...interface{}) {
+	loggers.write(newTLog(WarnLevel, tags, variables...))
+}
+
+// TWarnf will print a formatted message at warn level and meta-tag the log
+func TWarnf(tags Tags, format string, variables ...interface{}) {
+	loggers.write(newTLogf(WarnLevel, tags, format, variables...))
+}
+
 //
 // Info logging (Level 4)
 //
@@ -100,6 +130,16 @@ func Info(variables ...interface{}) {
 // Infof will print a formatted message to all loggers at info level
 func Infof(format string, variables ...interface{}) {
 	loggers.write(newLogf(InfoLevel, format, variables...))
+}
+
+// TInfo will print any number of variables at info level and meta-tag the log
+func TInfo(tags Tags, variables ...interface{}) {
+	loggers.write(newTLog(InfoLevel, tags, variables...))
+}
+
+// TInfof will print a formatted message at info level and meta-tag the log
+func TInfof(tags Tags, format string, variables ...interface{}) {
+	loggers.write(newTLogf(InfoLevel, tags, format, variables...))
 }
 
 //
@@ -116,6 +156,16 @@ func Debugf(format string, variables ...interface{}) {
 	loggers.write(newLogf(DebugLevel, format, variables...))
 }
 
+// TDebug will print any number of variables at debug level and meta-tag the log
+func TDebug(tags Tags, variables ...interface{}) {
+	loggers.write(newTLog(DebugLevel, tags, variables...))
+}
+
+// TDebugf will print a formatted message at debug level and meta-tag the log
+func TDebugf(tags Tags, format string, variables ...interface{}) {
+	loggers.write(newTLogf(DebugLevel, tags, format, variables...))
+}
+
 //
 // Trace logging (Level 6)
 //
@@ -128,6 +178,16 @@ func Trace(variables ...interface{}) {
 // Tracef will print a formatted message to all loggers at debug level
 func Tracef(format string, variables ...interface{}) {
 	loggers.write(newLogf(TraceLevel, format, variables...))
+}
+
+// TTrace will print any number of variables at trace level and meta-tag the log
+func TTrace(tags Tags, variables ...interface{}) {
+	loggers.write(newTLog(TraceLevel, tags, variables...))
+}
+
+// TTracef will print a formatted message at trace level and meta-tag the log
+func TTracef(tags Tags, format string, variables ...interface{}) {
+	loggers.write(newTLogf(TraceLevel, tags, format, variables...))
 }
 
 //
@@ -172,9 +232,16 @@ func SetColorLogging(colorLogging bool) {
 	}
 }
 
-// SetColorMap will loop through all the loggers and set the colors for each log level.
-func SetColorMap(colorMap ColorMap) {
+// SetLogLevelColorMap will loop through all the loggers and set the colors for each log level.
+func SetLogLevelColorMap(logLevelColorMap LogLevelColorMap) {
 	for _, logger := range loggers {
-		logger.SetColorMap(colorMap)
+		logger.SetLogLevelColorMap(logLevelColorMap)
+	}
+}
+
+// SetTagColorMap will loop through all the loggers and set the colors for each tag.
+func SetTagColorMap(tagColorMap TagColorMap) {
+	for _, logger := range loggers {
+		logger.SetTagColorMap(tagColorMap)
 	}
 }