+++ CHANGELOG.md
@@ -1,8 +1,27 @@
 # Changelog
 
+## v0.3.0
+
+- [#5] Added ability to add meta-tags to logs
+- Added custom tag color map support
+- Added examples:
+  - `tag_example` - Add meta-tags to logs
+
+**Changes:**
+
+- Examples:
+  - `color_example` - Added tag color example
+
+**Breaking Changes:**
+
+- Renamed `ColorMap` -> `InfoLevelColorMap` to distiguish it from `TagColorMap`
+  - All methods with the phrase `ColorMap` in them are also renamed.
+
 ## v0.2.0
 
-- [#3] Added custom color map support
+- [#3] Added custom log level color map support
+- Added examples:
+  - `color_example` - Change the default color map
 - Added some launch configurations for debugging
 
 ## v0.1.0